Fix the metrics reload when resizing the application
This reload happens because the widget tree changed (in SplitWiew widget) when the screen
width is lower/higher the defined breakpoint: a Row and an Expanded do not exist in the small
screen layout.

If Flutter is not able to find the same Widget (class) at the same position in the tree, then it considers
that it's a new widget and recreate it and its state if any.